feat: 그래프 배경 그리드를 그리기 위한 GraphBackgroundView 파일 분리
